feat: refactoring from script v1 to have config_parser v1 or v2
