+++ app/v1/modules/config_parser/__init__.py
@@ -1,0 +1,9 @@
+# -*- coding: utf-8 -*-
+
+
+from .script_config_parser import ScriptConfigParser
+
+
+__all__ = [
+    "ScriptConfigParser",
+]
